+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 **Changes in version 2.9**
 
+- Fixed County at DX algorithm setting the wrong input field.
+
+**Changes in version 2.9**
+
 - Removed unused input fields from County at DX Analysis algorithm. 
 - Fixed IARC algorithm crashing when input sequence number is blank.